+++ packages/server/lib/engine_io_server.dart
@@ -4,6 +4,7 @@ library engine_io_server;
 
 export 'package:engine_io_shared/exceptions.dart';
 export 'package:engine_io_shared/packets.dart';
+export 'package:engine_io_shared/socket.dart';
 export 'package:engine_io_shared/transports.dart' show ConnectionType;
 
 export 'src/transports/polling/polling.dart' show PollingTransport;
@@ -12,4 +13,3 @@ export 'src/transports/transport.dart' show Transport;
 export 'src/configuration.dart' show ConnectionOptions, ServerConfiguration;
 export 'src/server.dart' show Server;
 export 'src/socket.dart' show Socket;
-export 'src/upgrade.dart' show UpgradeStatus;