+++ packages/shared/lib/src/event_emitter.dart
@@ -1,7 +1,7 @@
 // ignore_for_file: one_member_abstracts
 
 /// Represents a class that stores and manages event streams.
-abstract interface class EventEmitter {
+abstract class EventEmitter {
   /// Closes all event streams.
   Future<void> closeStreams();
 }